+++ src/main/java/com/olap3/cubeexplorer/optimize/TimeableOp.java
@@ -1,8 +1,8 @@
 package com.olap3.cubeexplorer.optimize;
 
 public interface TimeableOp {
-    public int aprioriTime();
-    public int aposterioriTime();
+    public long aprioriTime();
+    public long aposterioriTime();
     public int aprioriTuples();
     public int aposterioriTuples();
     public double aprioriInterest();
